Update: Support not pass `folder_name` and choice `copy_image` and can set `output_name`

+++ script/convert_dataset_from_tmpco.py
@@ -22,7 +22,6 @@ Folder layout example:
 import argparse
 import logging
 import pprint
-import shutil
 from os import PathLike
 from pathlib import Path
 from typing import (
@@ -43,8 +42,11 @@ def arg_parser() -> argparse.Namespace:
 
     parser.add_argument("-r", "--root_path", type=str, required=True, help="Folder data path")
     parser.add_argument("-o", "--output_path", type=str, required=True, help="Output data path")
-    parser.add_argument("-f", "--folder_name", type=str, default="orig", help="Check folder name")
+    parser.add_argument("-f", "--folder_name", type=str, default=None, help="Check folder name")
     parser.add_argument("--format", type=str, default="latex", help="Output label format")
+    parser.add_argument("--output_name", type=str, default="data", help="Output filename")
+    parser.add_argument("--copy_image", action="store_true", help="Output filename")
+
     parser.add_argument("--tqdm", action="store_true", help="Show progress bar")
 
     parser.add_argument("-v", "--verbose", action="store_true", help="Show detail")
@@ -74,30 +76,31 @@ def check_iterate_num(
 def _read_label_data_and_copy_image(
     iterate_num: int,
     path_chain: list[PathLike],
-    folder_name: str,
+    folder_name: str = None,
 ) -> List[Dict[str, Any]]:
     root_path = "/".join([str(p) for p in path_chain])
     data = list()
 
     if len(path_chain) == iterate_num:
         read_filenames = set()
-        for path in Path(root_path, folder_name).glob(r"*.txt"):
-            if path.stem not in read_filenames:
-                base_name = path.stem
+        path = Path(root_path, folder_name) if folder_name else Path(root_path)
+        for filepath in path.glob(r"*.txt"):
+            if filepath.stem not in read_filenames:
+                base_name = filepath.stem
                 extension_name = None
                 logger.info(f"Read {base_name!s}")
                 read_filenames.add(base_name)
 
                 # Read image and check label file
-                if Path(root_path, folder_name, f"{base_name}.txt").exists():
-                    if Path(root_path, folder_name, f"{base_name}.jpg").exists():
+                if Path(path, f"{base_name}.txt").exists():
+                    if Path(path, f"{base_name}.jpg").exists():
                         extension_name = ".jpg"
-                    elif Path(root_path, folder_name, f"{base_name}.png").exists():
+                    elif Path(path, f"{base_name}.png").exists():
                         extension_name = ".png"
                     else:
                         raise FileNotFoundError("Not found image file")
 
-                    with Path(root_path, folder_name, f"{base_name}.txt").open("r", encoding="utf-8") as f:
+                    with Path(path, f"{base_name}.txt").open("r", encoding="utf-8") as f:
                         label_content = f.read()
                 else:
                     raise FileNotFoundError("Not found label file")
@@ -106,8 +109,8 @@ def _read_label_data_and_copy_image(
                     {
                         "label": label_content,
                         "source": base_name,
-                        "full_source": str(Path(root_path, folder_name, base_name)),
-                        "image_path": str(Path(root_path, folder_name, f"{base_name}{extension_name}")),
+                        "full_source": str(Path(path, base_name)),
+                        "image_path": str(Path(path, f"{base_name}{extension_name}")),
                     }
                 )
 
@@ -119,6 +122,7 @@ def _read_label_data_and_copy_image(
         data += _read_label_data_and_copy_image(
             iterate_num=iterate_num,
             path_chain=path_chain + [path.name],
+            folder_name=folder_name,
         )
     return data
 
@@ -126,8 +130,10 @@ def _read_label_data_and_copy_image(
 def convert_dataset_from_tmpco(
     root_path: PathLike,
     output_path: PathLike,
+    output_name: str,
     folder_name: str,
     format: str,
+    copy_image: bool = False,
     tqdm: bool = False,
 ) -> pd.DataFrame:
     # Pre-check
@@ -135,16 +141,17 @@ def convert_dataset_from_tmpco(
     output_path = Path(output_path)
     assert root_path.exists(), f"'{root_path!s}' not exist"
 
-    if output_path.exists():
+    if output_path.exists() and output_path.is_file():
         input(f"Warning: '{output_path!s}' is exists, will remove, if want press Enter...")
-        if output_path.is_dir():
-            shutil.rmtree(output_path)
-        else:
-            output_path.unlink()
+        output_path.unlink()
 
-    iterate_num = check_iterate_num(
-        path=root_path,
-        check_name=folder_name,
+    iterate_num = (
+        check_iterate_num(
+            path=root_path,
+            check_name=folder_name,
+        )
+        if folder_name
+        else 0
     )
     assert iterate_num is not None, f"Error: this '{root_path!s}' can't parse with tmpco folder layout"
     logger.info(f"Iterate count: {iterate_num}")
@@ -199,11 +206,12 @@ def convert_dataset_from_tmpco(
                 "full_source": df.iloc[i]["full_source"],
             }
         )
-        with Image.open(df.iloc[i]["image_path"]) as image_file:
-            image_file.save(str(Path(image_save_path, f'{df.iloc[i]["source"]}.jpg')))
+        if copy_image:
+            with Image.open(df.iloc[i]["image_path"]) as image_file:
+                image_file.save(str(Path(image_save_path, f'{df.iloc[i]["source"]}.jpg')))
     logger.info("Convert to dataset format end")
 
-    dataset_save_path = Path(output_path, "data.xlsx")
+    dataset_save_path = Path(output_path, Path(output_name).stem + ".xlsx")
     dataset = pd.DataFrame(dataset)
     logger.info("Dataset to pd.dataframe")
     dataset.to_excel(str(dataset_save_path))
